pyproject.toml
[tool.poetry]
name = "django-admin-pandasai"
version = "0.1.0"
description = ""
authors = ["Gonçalo Lopes <g.lopes@runtime-revolution.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
pyyaml = "^6.0.2"
django-environ = "^0.11.2"
numpy = "1.26.4"
pandasai = "^2.3.1"
django-adminfilters = "^2.5.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.28.0"
tqdm = "^4.66.5"
babel = "^2.16.0"
fastparquet = "^2024.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
